extract some utility functions into a module, mindmup_as_attack_trees
