+++ scripts/mindmup_as_attack_trees.py
@@ -1,0 +1,65 @@
+from collections import OrderedDict
+import re
+
+def get_node_children(node):
+	return OrderedDict(sorted(node.get('ideas', dict()).iteritems(), key=lambda t: float(t[0]))).values()
+
+def is_node_a_leaf(node):
+	return len(get_node_children(node)) == 0
+
+def get_node_title(node):
+	return node.get('title', '')
+
+def is_mitigation(node):
+	return is_node_a_leaf(node) and get_node_title(node).startswith('Mitigation:')
+
+def is_all_children(node, predicate):
+	for child in get_node_children(node):
+		if not predicate(child):
+		    return False
+	return True
+
+def is_attack_vector(node):
+	if is_node_a_leaf(node):
+	    return not is_mitigation(node)
+	else:
+	    return is_all_children(node, is_mitigation)
+
+def is_objective(node):
+	raw_description = get_raw_description(node)
+	return 'OBJECTIVE::' in raw_description
+
+def get_raw_description(node):
+	#prefer the mindmup 2.0 'note' to the 1.0 'attachment'
+	description = node.get('attr', dict()).get('note', dict()).get('text', '')
+	if description is '':
+		description = node.get('attr', dict()).get('attachment', dict()).get('content', '')
+
+	return description
+
+def get_node_referent_title(node):
+	title = node.get('title', '')
+
+	if '(*)' in node.get('title'):
+		wip_referent_title = title.replace('(*)','').strip()
+	else:
+		referent_coords = re.search(r'\((\d+\..*?)\)', title).groups()[0]
+		wip_referent_title = "%s %s" % (referent_coords, re.sub(r'\(\d+\..*?\)', '', title).strip())
+	return wip_referent_title
+
+def is_node_a_reference(node):
+	title = node.get('title', '')
+
+	return (not title.find('(*)') == -1) or (not re.search(r'\(\d+\..*?\)', title) is None)
+
+def get_node_referent(node, nodes_lookup):
+	node_referent_title = get_node_referent_title(node)
+	node_referent = nodes_lookup.get(node_referent_title, None)
+
+	if node_referent is None:
+		print("ERROR missing node referent: %s" % node_referent_title)
+		return node
+	else:
+		return node_referent
+
+
